refactor: update interface imports
