+++ src/adapters/api-keys/index.ts
@@ -1,3 +1,3 @@
-export * from './interfaces';
+export * from '@/core/api-keys/IApiKeyDataProvider';
 export * from './factory';
 export * from './supabase-adapter';
