+++ src/adapters/audit/index.ts
@@ -1,3 +1,3 @@
-export * from './interfaces';
+export * from '@/core/audit/IAuditDataProvider';
 export * from './factory';
 export * from './supabase-adapter';
