--- src/adapters/audit/interfaces.ts
@@ -1,7 +1,0 @@
-import { AuditLogEntry, AuditLogQuery } from '@/core/audit/models';
-
-export interface AuditDataProvider {
-  getUserActionLogs(
-    query: AuditLogQuery
-  ): Promise<{ logs: AuditLogEntry[]; count: number }>;
-}
