+++ src/adapters/csrf/index.ts
@@ -1,3 +1,3 @@
-export * from './interfaces';
+export * from '@/core/csrf/ICsrfDataProvider';
 export * from './factory';
 export * from './default-adapter';
