--- src/adapters/csrf/interfaces.ts
@@ -1,3 +1,0 @@
-export interface CsrfDataProvider {
-  generateToken(): Promise<string>;
-}
