+++ src/adapters/notification/index.ts
@@ -1,1 +1,1 @@
-export * from './interfaces';
+export * from '@/core/notification/INotificationDataProvider';
