+++ src/adapters/permission/index.ts
@@ -4,6 +4,6 @@
  * This file exports all permission adapter related types and functions.
  */
 
-export * from './interfaces';
+export * from '@/core/permission/IPermissionDataProvider';
 export * from './factory';
 export * from './supabase-permission-provider';
