--- src/adapters/permission/interfaces.ts
@@ -1,1 +1,0 @@
-export { IPermissionDataProvider as PermissionDataProvider } from "../../core/permission/IPermissionDataProvider";
