+++ src/adapters/session/index.ts
@@ -1,3 +1,3 @@
-export * from './interfaces';
+export * from '@/core/session/ISessionDataProvider';
 export * from './factory';
 export * from './supabase-adapter';
