+++ src/adapters/sso/index.ts
@@ -1,3 +1,3 @@
-export * from './interfaces';
+export * from '@/core/sso/ISsoDataProvider';
 export * from './factory';
 export * from './supabase-adapter';
