+++ src/adapters/subscription/index.ts
@@ -1,3 +1,3 @@
-export * from './interfaces';
+export * from '@/core/subscription/ISubscriptionDataProvider';
 export * from './factory';
 export * from './supabase-adapter';
