+++ src/adapters/team/index.ts
@@ -4,6 +4,6 @@
  * This file exports all team adapter related types and functions.
  */
 
-export * from './interfaces';
+export * from '@/core/team/ITeamDataProvider';
 export * from './factory';
 export * from './supabase-team-provider';
