+++ src/adapters/user/index.ts
@@ -4,6 +4,6 @@
  * This file exports all user adapter related types and functions.
  */
 
-export * from './interfaces';
+export * from '@/core/user/IUserDataProvider';
 export * from './factory';
 export * from './supabase-user-provider';
