--- src/adapters/user/interfaces/index.ts
@@ -1,9 +1,0 @@
-/**
- * User Adapter Interfaces
- * 
- * This file exports all user-related adapter interfaces.
- */
-
-export * from './user-data-provider';
-
-// Add other user-related adapter interfaces here as needed
