+++ src/services/csrf/default-csrf.service.ts
@@ -1,6 +1,6 @@
 import { CsrfService } from '@/core/csrf/interfaces';
 import { CsrfToken } from '@/core/csrf/models';
-import type { CsrfDataProvider } from '@/adapters/csrf/interfaces';
+import type { CsrfDataProvider } from '@/core/csrf/ICsrfDataProvider';
 
 export class DefaultCsrfService implements CsrfService {
   constructor(private csrfProvider: CsrfDataProvider) {}
