+++ src/services/webhooks/index.ts
@@ -1,6 +1,6 @@
 import type { AxiosInstance } from 'axios';
 import type { IWebhookService } from '@/core/webhooks/IWebhookService';
-import type { IWebhookDataProvider } from '@/adapters/webhook/IWebhookDataProvider';
+import type { IWebhookDataProvider } from '@/core/webhooks/IWebhookDataProvider';
 import { WebhookService } from './WebhookService';
 
 export interface WebhookServiceConfig {
